Removing old api code from batch-email-sender.
